LC-count-nice-pairs-in-an-array [Runtime: N/A; Memory: N/A]

+++ array/count-nice-pairs-in-an-array.java
@@ -27,6 +27,6 @@ class Solution {
             size--;
         }
         
-        return revNum % (Math.pow(10,9)+7);
+        return revNum % (int)(Math.pow(10,9)+7);
     }
 }
